Implement signup and login functionality.

+++ Final Servlet Project/ItemsApp/src/main/java/model/User.java
@@ -2,24 +2,24 @@ package model;
 
 public class User {
 
-    private int id;
+    private Long id;
     private String username;
     private String password;
 
     public User() {
     }
 
-    public User(int id, String username, String password) {
+    public User(Long id, String username, String password) {
         this.id = id;
         this.username = username;
         this.password = password;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
